+++ src/main/java/xfactor/Arrayer.java
@@ -1,0 +1,60 @@
+package xfactor;
+
+public class Arrayer {
+
+    // summation() accepts an array of integers and returns the sum of all numbers within.
+    // ex: [0, 2, -1, 15] -> 16
+    public static int summation(int[] numbers) {
+        // TODO implement
+        return 0;
+    }
+
+    // findNegative() accepts an array of integers containing one negative number
+    // and returns the *position* of the negative number. You may safely assume
+    // the provided array contains only a single negative number.
+    // ex: [11, 13, -1, 0, 9] -> 2
+    public static int findNegative(int[] numbers) {
+        // TODO implement
+        return 0;
+    }
+
+    // remove() accepts an array of integers and an int n. The method removes *all instances*
+    // of n from the provided array and returns a new array with no instances of n.
+    // ex: [0, 1, 1, 2, 2, 3], n = 2 -> [0, 1, 1, 3]
+    public static int[] remove(int[] numbers, int n) {
+        // TODO implement
+        return new int[0];
+    }
+
+    // merge() accepts two *pre-sorted* arrays of integers and returns a new *sorted* array.
+    // You may safely assume that the provided arrays are always of equal length.
+    // ex: [0, 2, 4, 8] + [1, 3, 5, 7] -> [0, 1, 2, 3, 4, 5, 7, 8]
+    public static int[] merge(int[] first, int[] second) {
+        // TODO implement
+        return new int[0];
+    }
+
+    // roundUp() accepts an array of *non-negative* doubles and returns an array of
+    // rounded integers. Doubles are rounded up when the decimal is >= 0.5.
+    // ex: [1.2, 3.5, 4.2, 0.0] -> [1, 4, 4, 0]
+    public static int[] roundUp(double[] doubles) {
+        // TODO implement
+        return new int[0];
+    }
+
+    // evensOnly() accepts an array of integers and returns a new array containing
+    // only the even numbers found in the provided array, in their original order.
+    // ex: [3, 4, 7, 8, 12] -> [4, 8, 12]
+    public static int[] evensOnly(int[] numbers) {
+        // TODO implement
+        return new int[0];
+    }
+
+    // lastOfFourDigits() accepts an array of four-digit integers and returns a new
+    // array containing only the last digit of each number in the original sequence.
+    // ex: [1004, 1112, 5667, 8009] -> [4, 2, 7, 9]
+    public static int[] lastOfFourDigits(int[] numbers) {
+        // TODO implement
+        return new int[0];
+    }
+}
