+++ src/main/java/xfactor/Fibonacci.java
@@ -2,7 +2,13 @@ package xfactor;
 
 public class Fibonacci {
     
+    // sequence() accepts an integer n and returns the value of
+    // the nth position of a standard fibonacci sequence. Recall
+    // the fibonacci sequence is defined as follows ...
+    // 0, 1, 1, 2, 3, 5, 8 ...
+    // ex: n = 6 -> 8
     public static int sequence(int n) {
+        // TODO implement
         return 0;
     }
 }
