--- src/main/java/xfactor/XFactor.java
@@ -1,32 +1,0 @@
-package xfactor;
-
-public class XFactor {
-
-    public static int summation(int[] numbers) {
-        return 0;
-    }
-
-    public static int findNegative(int[] numbers) {
-        return 0;
-    }
-
-    public static int[] remove(int[] numbers, int n) {
-        return new int[0];
-    }
-
-    public static int[] merge(int[] first, int[] second) {
-        return new int[0];
-    }
-
-    public static int[] roundUp(double[] doubles) {
-        return new int[0];
-    }
-
-    public static int[] evensOnly(int[] numbers) {
-        return new int[0];
-    }
-
-    public static int[] threeOfFourDigits(int[] numbers) {
-        return new int[0];
-    }
-}
